+++ src/test/java/subway/validator/StationNameValidatorTest.java
@@ -1,22 +1,11 @@
-package subway.view.validator;
+package subway.validator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class InputValidatorTest {
+public class StationNameValidatorTest {
 
-    private final InputValidator validator = new InputValidator();
-
-    @Test
-    @DisplayName("입력 값이 null 일 경우 예외 발생")
-    public void checkNull_Null_ExceptionThrown() {
-
-        // given
-        String input = null;
-
-        // when, then
-        InputValidatorUtils.assertValidationFailure(input, validator, InputValidator.NULL_ERROR);
-    }
+    private final Validator validator = new StationNameValidator();
 
     @Test
     @DisplayName("입력 값의 길이가 최소 길이 미만일 경우 예외 발생")
@@ -26,8 +15,8 @@ public class InputValidatorTest {
         String input = "역";
 
         // when, then
-        InputValidatorUtils.assertValidationFailure(input, validator, InputValidator.RANGE_ERROR +
-                String.format(InputValidator.INPUT_LENGTH_MESSAGE, input.length()));
+        ValidatorUtils.assertValidationFailure(input, validator, StationNameValidator.RANGE_ERROR +
+                String.format(StationNameValidator.INPUT_LENGTH_MESSAGE, input.length()));
     }
 
     @Test
@@ -38,8 +27,8 @@ public class InputValidatorTest {
         String input = "서울특별시강남구강남역";
 
         // when, then
-        InputValidatorUtils.assertValidationFailure(input, validator, InputValidator.RANGE_ERROR +
-                String.format(InputValidator.INPUT_LENGTH_MESSAGE, input.length()));
+        ValidatorUtils.assertValidationFailure(input, validator, StationNameValidator.RANGE_ERROR +
+                String.format(StationNameValidator.INPUT_LENGTH_MESSAGE, input.length()));
     }
 
     @Test
@@ -50,8 +39,8 @@ public class InputValidatorTest {
         String input = "서울시양재동시민의숲";
 
         // when, then
-        InputValidatorUtils.assertValidationFailure(input, validator, InputValidator.RANGE_ERROR +
-                String.format(InputValidator.INPUT_LENGTH_MESSAGE, input.length() + 1));
+        ValidatorUtils.assertValidationFailure(input, validator, StationNameValidator.RANGE_ERROR +
+                String.format(StationNameValidator.INPUT_LENGTH_MESSAGE, input.length() + 1));
     }
 
     @Test
@@ -62,7 +51,7 @@ public class InputValidatorTest {
         String input = "abc";
 
         // when, then
-        InputValidatorUtils
-                .assertValidationFailure(input, validator, InputValidator.NOT_KOREAN_ERROR);
+        ValidatorUtils
+                .assertValidationFailure(input, validator, StationNameValidator.NOT_KOREAN_ERROR);
     }
 }
